Fixed dependency tests, added exceptions.
