+++ geode-core/src/test/java/org/apache/geode/distributed/internal/membership/gms/api/MembershipAPIArchUnitTest.java
@@ -28,10 +28,14 @@ import org.junit.runner.RunWith;
 import org.apache.geode.distributed.DistributedMember;
 import org.apache.geode.distributed.internal.ClusterDistributionManager;
 import org.apache.geode.distributed.internal.DistributionMessage;
+import org.apache.geode.distributed.internal.InternalDistributedSystem;
 import org.apache.geode.distributed.internal.membership.InternalDistributedMember;
+import org.apache.geode.distributed.internal.membership.MembershipTestHook;
 import org.apache.geode.distributed.internal.membership.MembershipView;
+import org.apache.geode.distributed.internal.membership.QuorumChecker;
 import org.apache.geode.distributed.internal.membership.gms.MemberDataBuilderImpl;
 import org.apache.geode.distributed.internal.membership.gms.MembershipBuilderImpl;
+import org.apache.geode.distributed.internal.membership.gms.Services;
 
 @RunWith(ArchUnitRunner.class)
 @AnalyzeClasses(packages = "org.apache.geode.distributed.internal.membership.gms.api")
@@ -58,5 +62,11 @@ public class MembershipAPIArchUnitTest {
               .or(type(DistributedMember.class))
               .or(type(InternalDistributedMember[].class))
               .or(type(DistributionMessage.class))
-              .or(type(ClusterDistributionManager.class)));
+              .or(type(ClusterDistributionManager.class))
+
+              // TODO: GMSMembershipManager move brought these...
+              .or(type(QuorumChecker.class))
+              .or(type(MembershipTestHook.class))
+              .or(type(InternalDistributedSystem.class))
+              .or(type(Services.class)));
 }
